+++ test/Test/Runtime/EndToEndAutogradGPU.hs
@@ -113,6 +113,45 @@ tests getGPU = testGroup "EndToEnd.AutogradGPU"
         let expected = V.fromList [1, 2, 1, 2, 4, 2, 1, 2, 1]
         assertBool "conv2d grad close" $
             V.and (V.zipWith (\r e -> abs (r - e) < 0.01) result expected)
+    , testCase "grad slice stride" $ do
+        GPUResource api client dev <- getGPU
+        let f x = do
+                y <- slice @'[5] @'[3] x (v1 0) (v1 5) (v1 2)
+                sumAll y
+            gradModu = gradModule @'[5] @'F32 f
+        exec <- compile api client (render gradModu)
+        let inp = V.fromList [1.0, 2.0, 3.0, 4.0, 5.0]
+        bufIn <- toDeviceF32On api client dev inp [5]
+        [bufOut] <- executeOn api exec dev [bufIn]
+        result <- fromDeviceF32 api bufOut 5
+        let expected = V.fromList [1.0, 0.0, 1.0, 0.0, 1.0]
+        result @?= expected
+    , testCase "grad conv2d stride" $ do
+        GPUResource api client dev <- getGPU
+        let f x = do
+                k <- constant @'[3, 3, 1, 1] @'F32 1.0
+                y <- conv2dWithPadding @1 @4 @4 @1 @1 @3 @3 @2 @2 (v2 2 2) (p2 (1,1) (1,1)) x k
+                sumAll y
+            gradModu = gradModule @'[1, 4, 4, 1] @'F32 f
+        exec <- compile api client (render gradModu)
+        let inp = V.fromList [1.0..16.0]
+        bufIn <- toDeviceF32On api client dev inp [1, 4, 4, 1]
+        [bufOut] <- executeOn api exec dev [bufIn]
+        result <- fromDeviceF32 api bufOut 16
+        assertBool "grad non-zero" $ V.sum result > 0
+    , testCase "grad transposeConvolution asymmetric pad" $ do
+        GPUResource api client dev <- getGPU
+        let f x = do
+                k <- constant @'[3, 3, 1, 1] @'F32 1.0
+                y <- transposeConvolution @1 @2 @2 @1 @1 @3 @3 @2 @2 (v2 2 2) (p2 (1,0) (1,0)) x k
+                sumAll y
+            gradModu = gradModule @'[1, 2, 2, 1] @'F32 f
+        exec <- compile api client (render gradModu)
+        let inp = V.fromList [1.0, 2.0, 3.0, 4.0]
+        bufIn <- toDeviceF32On api client dev inp [1, 2, 2, 1]
+        [bufOut] <- executeOn api exec dev [bufIn]
+        result <- fromDeviceF32 api bufOut 4
+        assertBool "grad non-zero" $ V.sum result > 0
     , testCase "grad maxPool" $ do
         GPUResource api client dev <- getGPU
         let f x = do
@@ -130,6 +169,20 @@ tests getGPU = testGroup "EndToEnd.AutogradGPU"
         let expected = V.fromList [0,0,0,0, 0,1,0,1, 0,0,0,0, 0,1,0,1]
         assertBool "maxPool grad close" $
             V.and (V.zipWith (\r e -> abs (r - e) < 0.01) result expected)
+    , testCase "grad pad interior" $ do
+        GPUResource api client dev <- getGPU
+        let f x = do
+                padVal <- constant @'[] @'F32 0.0
+                y <- pad @'[2] @'[3] x padVal (v1 0) (v1 0) (v1 1)
+                sumAll y
+            gradModu = gradModule @'[2] @'F32 f
+        exec <- compile api client (render gradModu)
+        let inp = V.fromList [1.0, 2.0]
+        bufIn <- toDeviceF32On api client dev inp [2]
+        [bufOut] <- executeOn api exec dev [bufIn]
+        result <- fromDeviceF32 api bufOut 2
+        let expected = V.fromList [1.0, 1.0]
+        result @?= expected
     , testCase "grad concatenate2" $ do
         GPUResource api client dev <- getGPU
         let f a b = do